requirements.txt
Flask>=2.0
tensorflow>=2.9
Pillow
numpy
werkzeug
matplotlib
